+++ CHANGELOG.md
@@ -9,8 +9,8 @@ Ability to add/change hyperlinks anchored to bookmarks/headings
 * 
 
 ### Spreadsheet Editor
-* The following functions are added: GetRows, GetCols, GetCount
-* The following properties are added: Rows, Cols, Count
+* The following functions are added: GetRows, GetCols, GetCount, GetHidden, SetHidden
+* The following properties are added: Rows, Cols, Count, Hidden
 
 ### Presentation Editor
 *